ZmBlockImport completed. Block page still displays mev when no zm data, and controls appropriately disabled/enabled

+++ ZeroMev/MevEFC/Price.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace ZeroMev.MevEFC
 {
     public partial class Price
     {
         public DateTime Timestamp { get; set; }
-        public decimal UsdPrice { get; set; }
+        public BigInteger UsdPrice { get; set; }
         public string TokenAddress { get; set; } = null!;
     }
 }
